+++ openneuro/__init__.py
@@ -1,8 +1,4 @@
-try:
-    from importlib import metadata
-except ImportError:
-    # Running on pre-3.8 Python; use importlib-metadata package
-    import importlib_metadata as metadata
+from importlib import metadata
 
 try:
     __version__ = metadata.version("openneuro-py")