+++ openneuro/_download.py
@@ -7,11 +7,7 @@ import asyncio
 from pathlib import Path, PurePosixPath
 import string
 import json
-from typing import Optional, Union
-if sys.version_info >= (3, 8):
-    from typing import Literal
-else:
-    from typing_extensions import Literal
+from typing import Optional, Union, Literal
 
 if sys.version_info >= (3, 9):
     from collections.abc import Iterable, Generator